+++ lib/constants_generator/constants_generator.dart
@@ -1,4 +1,4 @@
-library binary_generator.library_generator;
+library binary_generator.constants_generator;
 
 import "package:binary_declarations/binary_declarations.dart";
 import "package:binary_generator/internal/generators/generators.dart";
@@ -9,8 +9,7 @@ import "package:macro_processor/macro_expander.dart";
 import "package:template_block/template_block.dart";
 
 part 'src/class_library_generator.dart';
+part 'src/constants_generator.dart';
 part 'src/constructor_generator.dart';
-part 'src/foreign_function_generator.dart';
 part 'src/getter_type_generator.dart';
-part 'src/library_generator.dart';
 part 'src/script_generator.dart';